requirements.txt
numpy
ase
torch
e3nn
hydra-core
hydra-submitit-launcher
tensorboard
pandas
scipy
lz4
boto3